+++ src/main/java/edu/architect_711/wordsapp/controller/LanguageController.java
@@ -1,6 +1,6 @@
 package edu.architect_711.wordsapp.controller;
 
-import edu.architect_711.wordsapp.model.dto.LanguageDto;
+import edu.architect_711.wordsapp.model.dto.language.LanguageDto;
 import edu.architect_711.wordsapp.service.language.LanguageService;
 import io.swagger.v3.oas.annotations.security.SecurityRequirements;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +10,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-@RestController @RequestMapping("/api/languages")
+@RestController
+@RequestMapping("/api/languages")
 @RequiredArgsConstructor
 public class LanguageController {
     private final LanguageService languageService;
