+++ src/main/java/edu/architect_711/wordsapp/model/entity/Language.java
@@ -6,11 +6,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Entity @Table(name = "language")
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor
+@Entity
+@Table(name = "language")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class Language {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false, unique = true)
