+++ src/main/java/edu/architect_711/wordsapp/service/group/DefaultGroupService.java
@@ -1,11 +1,10 @@
 package edu.architect_711.wordsapp.service.group;
 
-import edu.architect_711.wordsapp.model.dto.GroupDto;
-import edu.architect_711.wordsapp.model.dto.SaveGroupDto;
-import edu.architect_711.wordsapp.model.dto.UpdateGroupDto;
+import edu.architect_711.wordsapp.model.dto.group.GroupDto;
+import edu.architect_711.wordsapp.model.dto.group.SaveGroupDto;
+import edu.architect_711.wordsapp.model.dto.group.UpdateGroupDto;
 import edu.architect_711.wordsapp.model.entity.Account;
 import edu.architect_711.wordsapp.model.entity.Group;
-import edu.architect_711.wordsapp.model.mapper.GroupMapper;
 import edu.architect_711.wordsapp.repository.GroupRepository;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +17,7 @@ import java.util.Objects;
 
 import static edu.architect_711.wordsapp.model.mapper.GroupMapper.toDto;
 import static edu.architect_711.wordsapp.model.mapper.GroupMapper.toEntity;
-import static edu.architect_711.wordsapp.security.AuthenticationExtractor.getAccount;
+import static edu.architect_711.wordsapp.security.AuthenticatedUserExtractor.getAccount;
 
 @Service
 @RequiredArgsConstructor
@@ -36,18 +35,14 @@ public class DefaultGroupService implements GroupService {
 
     @Override
     public List<GroupDto> get() {
-        return groupRepository
-            .findAll(getAccount().getId())
-            .stream()
-            .map(GroupMapper::toDto)
-            .toList();
+        return toDto(groupRepository.findAll(getAccount().getId()));
     }
 
     @Override
     public GroupDto update(@Valid UpdateGroupDto updateGroupDto) {
         Group group = groupRepository.safeFindById(updateGroupDto.getId());
 
-        checkGroupId(group.getAccount().getId());
+        checkThirdPartyAccess(group.getAccount().getId());
 
         group.setTitle(updateGroupDto.getTitle());
         group.setDescription(updateGroupDto.getDescription());
@@ -58,12 +53,17 @@ public class DefaultGroupService implements GroupService {
     @Override
     public void delete(Long id) {
         var group = groupRepository.safeFindById(id);
-        checkGroupId(group.getAccount().getId());
+        checkThirdPartyAccess(group.getAccount().getId());
 
         groupRepository.deleteById(id);
     }
 
-    private void checkGroupId(long groupAccountId) {
+    /**
+     * Check third-party access. For example when a user1 tries to access
+     * a user2 groups
+     * @param groupAccountId the group's owner id to be affected
+     */
+    private void checkThirdPartyAccess(long groupAccountId) {
         var account = getAccount();
 
         if (!Objects.equals(groupAccountId, account.getId()))
