+++ src/main/java/edu/architect_711/wordsapp/service/language/DefaultLanguageService.java
@@ -1,12 +1,13 @@
 package edu.architect_711.wordsapp.service.language;
 
-import edu.architect_711.wordsapp.model.dto.LanguageDto;
-import edu.architect_711.wordsapp.model.mapper.LanguageMapper;
+import edu.architect_711.wordsapp.model.dto.language.LanguageDto;
 import edu.architect_711.wordsapp.repository.LanguageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static edu.architect_711.wordsapp.model.mapper.LanguageMapper.toDto;
 
 @Service @RequiredArgsConstructor
 public class DefaultLanguageService implements LanguageService {
@@ -14,6 +15,6 @@ public class DefaultLanguageService implements LanguageService {
 
     @Override
     public List<LanguageDto> get() {
-        return languageRepository.findAll().stream().map(LanguageMapper::toDto).toList();
+        return toDto(languageRepository.findAll());
     }
 }
