+++ src/test/java/edu/architect_711/wordsapp/service/LanguageServiceIntegrationTest.java
@@ -1,6 +1,6 @@
 package edu.architect_711.wordsapp.service;
 
-import edu.architect_711.wordsapp.model.dto.LanguageDto;
+import edu.architect_711.wordsapp.model.dto.language.LanguageDto;
 import edu.architect_711.wordsapp.service.language.LanguageService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
